Port AirSLAM to ROS 2 Jazzy + TensorRT 10 (phases 0-2 complete)
Build system:
- package.xml format 3 (ament_cmake) replacing catkin
- CMakeLists.txt: ament + two-lib split (air_slam_core_lib non-ROS,
  air_slam_lib ROS-aware)
- CMakeLists.txt.ros1.bak preserves the original for reference

TensorRT 8 -> TensorRT 10 migration:
- 3rdparty/tensorrtbuffer/include/buffers.h: BufferManager keyed by
  tensor name (getNbIOTensors / setInputTensorAddress / enqueueV3),
  mDeviceBindings is now std::map<string,void*>, new helper
  setTensorAddresses(context). Compile-only smoke test added at
  3rdparty/tensorrtbuffer/test/buffers_compile_test.cpp.
- safe_common.h: roundUp deduction fix for TRT 10 Dims.d[] = int64_t.
- sample_entrypoints.h: drop NvCaffeParser.h / NvUffParser.h (parsers
  removed in TRT 10).
- super_point.cpp / super_glue.cpp / light_glue.cpp / plnet.cpp:
  setBindingDimensions -> setInputShape, executeV2(bindings) ->
  setTensorAddresses + enqueueV3 + cudaStreamSynchronize. Each network
  owns a private cudaStream_t. PLNet's cached binding-index members
  removed — TRT 10 keys directly by tensor name.

ROS 1 -> ROS 2 Jazzy port:
- ros_publisher.{h,cc}: ros::Publisher -> rclcpp::Publisher<T>::SharedPtr,
  tf::TransformBroadcaster -> tf2_ros::TransformBroadcaster member.
  Constructor takes rclcpp::Node::SharedPtr.
- MapBuilder / MapRefiner / MapUser constructor sigs updated to
  rclcpp::Node::SharedPtr.
- 4 demos (visual_odometry, map_refinement, relocalization,
  test_feature) rewritten with rclcpp::init / declare_parameter /
  rclcpp::ok / rclcpp::shutdown.
- 10 ROS 1 .launch XMLs converted to ROS 2 .launch.py.

utils.h decoupling:
- g2o::Line3D-dependent helpers extracted to new include/utils_g2o.h
  so the network/feature path compiles without g2o on the include path.
- mapline / map / line_processor / g2o_optimization headers now pull
  utils_g2o.h explicitly when they need Line3D.

Phase 3 prep:
- scripts/build_engines.sh regenerates all 5 .engine files from the
  bundled ONNX models using TRT 10 trtexec (FP32 first, then --fp16).
- README_JAZZY.md documents the port deltas + the manual apt install
  step (libg2o-dev, libgoogle-glog-dev, libgflags-dev) needed before
  the rest of the .cc files can link.

Verified green: colcon build --packages-select air_slam --symlink-install
produces libair_slam_core_lib.so + libair_slam_lib.so + the smoke test
on Ubuntu 24.04 + Jazzy + TensorRT 10.16 + CUDA 12.6 + RTX 4070.

+++ 3rdparty/tensorrtbuffer/include/safe_common.h
@@ -110,7 +110,8 @@ namespace tensorrt_buffer
     {
         if (vecDim >= 0)
         {
-            dims.d[vecDim] = roundUp(dims.d[vecDim], comps);
+            // TRT 10: dims.d is int64_t — explicit cast so roundUp<T> deduces.
+            dims.d[vecDim] = roundUp(dims.d[vecDim], static_cast<int64_t>(comps));
         }
         return tensorrt_buffer::volume(dims) * std::max(batch, 1);
     }
